requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
polars[deltalake]==0.20.5
deltalake
azure-storage-blob
adlfs